Move gqlCalls to apiCalls folder
